requirements.txt
Flask>=2.0.3,<3.0.0
# TODO Rever compatibilidade abaixo (as versões além da 2 não acitam o como passamos parâmetros, mas, em nossa imagem, não quebra).
# SQLAlchemy>=1.4.32,<2.0.0
SQLAlchemy>=1.4.32,<3.0.0
pg8000>=1.24.1,<2.0.0
sqlparams>=3.0.0,<4.0.0
pydantic>=1.9.0,<3.0.0
python-dateutil>=2.9.0,<3.0.0
PyMySQL>=1.0.2,<2.0.0
py-healthcheck>=1.10.1,<2.0.0
unidecode>=1.0.22,<2.0.0
opentelemetry-exporter-otlp>=1.32.1,<2.0.0
pyyaml>=6.0.3,<7.0.0
